+++ lib/super/form/builder.rb
@@ -71,7 +71,8 @@ module Super
           )
           options[:class] = join_classes(defaults[:class], options[:class])
           options[:data] = defaults[:data].deep_merge(options[:data] || {})
-          options[:value] = @builder.object.public_send(attribute).presence&.iso8601
+          options[:value] = @builder.object.public_send(attribute).presence
+          options[:value] = options[:value].iso8601 if options[:value].respond_to?(:iso8601)
 
           @builder.text_field(attribute, options)
         end
@@ -91,7 +92,8 @@ module Super
           )
           options[:class] = join_classes(defaults[:class], options[:class])
           options[:data] = defaults[:data].deep_merge(options[:data] || {})
-          options[:value] = @builder.object.public_send(attribute).presence&.iso8601
+          options[:value] = @builder.object.public_send(attribute).presence
+          options[:value] = options[:value].iso8601 if options[:value].respond_to?(:iso8601)
 
           @builder.text_field(attribute, options)
         end
@@ -112,7 +114,8 @@ module Super
           )
           options[:class] = join_classes(defaults[:class], options[:class])
           options[:data] = defaults[:data].deep_merge(options[:data] || {})
-          options[:value] = @builder.object.public_send(attribute).presence&.strftime("%H:%M:%S")
+          options[:value] = @builder.object.public_send(attribute).presence
+          options[:value] = options[:value].strftime("%H:%M:%S") if options[:value].respond_to?(:strftime)
 
           @builder.text_field(attribute, options)
         end